- Fixed a typing issue

+++ voxelmorph_model.py
@@ -456,7 +456,7 @@ class VoxelMorph(nn.Module):
 
         source_tensor = to_torch_var(source)
         target_tensor = to_torch_var(target)
-        mask_tensor = to_torch_var(mask)
+        mask_tensor = to_torch_var(mask.astype(np.int32))
 
         with torch.no_grad():
             input_s = torch.cat([source_tensor, target_tensor], dim=1)
